Enabling Docker deployment persistence (#20)
* Now using placement constraints.

* Added/improved scripts for cleaning up Docker deployments.

* Now waiting for ActiveMQ service.

* Now using $MPF_HOME/share/config path.

+++ scripts/docker-swarm-cleanup.sh
@@ -32,22 +32,131 @@
 
 printUsage() {
   echo "Usages:"
-  echo "docker-swarm-cleanup.sh [--ask-pass]"
+  echo "docker-swarm-cleanup.sh [--ask-pass] <--mysql-volume|--all-volumes|--no-volumes> [--remove-shared-data]"
   exit -1
 }
 
-if [ $# -eq 0 ]; then
-  askPass=0
-elif [ $# -eq 1 ]; then
-  if [ "$1" != "--ask-pass" ]; then
+# parseVolumeType(1: volumeType)
+parseVolumeType() {
+  if [ "$1" == "--mysql-volume" ]; then
+    removeMysqlVolume=1
+  elif [ "$1" == "--all-volumes" ]; then
+    removeAllVolumes=1
+  elif [ "$1" != "--no-volumes" ]; then
     printUsage
   fi
-  askPass=1
+}
+
+# parseRemoveSharedData(1: option)
+parseRemoveSharedData() {
+  if [ "$1" == "--remove-shared-data" ]; then
+    removeSharedData=1
+  else
+    printUsage
+  fi
+}
+
+# cleanupContainers(1: sshCmd)
+cleanupContainers() {
+sshCmd="$1"
+"${sshCmd[@]}" << "EOF"
+containerIds=$(docker ps -a -f name=openmpf_ -q)
+if [ ! -z "$containerIds" ]; then
+  echo "Removing openmpf containers:"
+  docker container rm -f $containerIds;
+else \
+  echo "No openmpf containers running."
+fi
+echo
+exit
+EOF
+}
+
+# cleanupMysqlVolume(1: sshCmd)
+cleanupMysqlVolume() {
+sshCmd="$1"
+"${sshCmd[@]}" << "EOF"
+# Don't remove openmpf_shared_data.
+volumeIds=$(docker volume ls -f name=openmpf_mysql_data -q)
+if [ ! -z "$volumeIds" ]; then
+  echo "Removing openmpf volumes:"
+  docker volume rm -f $volumeIds;
+else
+  echo "No openmpf volumes found."
+fi
+echo
+exit
+EOF
+}
+
+# cleanupAllVolumes(1: sshCmd)
+cleanupAllVolumes() {
+sshCmd="$1"
+"${sshCmd[@]}" << "EOF"
+volumeIds=$(docker volume ls -f name=openmpf -q)
+if [ ! -z "$volumeIds" ]; then
+  echo "Removing openmpf volumes:"
+  docker volume rm -f $volumeIds;
+else
+  echo "No openmpf volumes found."
+fi
+echo
+exit
+EOF
+}
+
+# forAllNodes(1: nodeIds, 2: function)
+forAllNodes() {
+  while read -r nodeId; do
+    nodeAddr=$(docker node inspect "$nodeId" --format '{{ .Status.Addr }}')
+    echo
+    echo "Connecting to $nodeAddr ..."
+
+    if [ "$askPass" = 0 ]; then
+      sshCmd=(ssh -oStrictHostKeyChecking=no "$nodeAddr" /bin/bash)
+    else
+      sshCmd=(sshpass -p "$sshPass" ssh -oStrictHostKeyChecking=no "$sshUser"@"$nodeAddr" /bin/bash)
+    fi
+
+    $2 "$sshCmd"
+  done <<< "$1"
+}
+
+askPass=0
+removeMysqlVolume=0
+removeAllVolumes=0
+removeSharedData=0
+
+if [ $# -eq 1 ]; then
+  parseVolumeType "$1"
+elif [ $# -eq 2 ]; then
+  if [ "$1" == "--ask-pass" ]; then
+    askPass=1
+    parseVolumeType "$2"
+  else
+    parseVolumeType "$1"
+    parseRemoveSharedData "$2"
+  fi
+elif [ $# -eq 3 ]; then
+  if [ "$1" == "--ask-pass" ]; then
+    askPass=1
+  else
+    printUsage
+  fi
+  parseVolumeType "$2"
+  parseRemoveSharedData "$3"
 else
   printUsage
 fi
 
-if [ $askPass = 1 ]; then
+# Abort early if the helper script doesn't exist.
+scriptPath="$(dirname $0)/docker-swarm-clean-shared-volume.sh"
+if [ ! -f "$scriptPath" ]; then
+  echo "Could not find \"$scriptPath\". Aborting."
+  exit -1
+fi
+
+if [ "$askPass" == 1 ]; then
   read -s -p "Enter SSH user: " sshUser
   echo
 
@@ -56,58 +165,34 @@ if [ $askPass = 1 ]; then
   echo
 fi
 
-# The following command does not always remove the stopped containers;
-# however, it should remove networks.
-# Refer to https://github.com/moby/moby/issues/32620.
-docker stack rm openmpf
-echo
+# Check if stack exists. If so, remove it.
+if docker stack ps openmpf > /dev/null 2>&1; then
+  # The following command does not always remove the stopped containers.
+  # Refer to https://github.com/moby/moby/issues/32620.
+  # According to the official docs: "Services, networks, and secrets associated with the stack will be removed."
+  docker stack rm openmpf
 
-# Give the previous command time to work.
-sleep 10
+  # Give the previous command time to work.
+  sleep 10
+fi
 
 nodeIds=$(docker node ls | sed -n '1!p' | cut -d ' ' -f 1)
 
-while read -r nodeId; do
-  nodeAddr=$(docker node inspect "$nodeId" --format '{{ .Status.Addr }}')
-  echo
-  echo "Connecting to $nodeAddr ..."
+forAllNodes "$nodeIds" cleanupContainers
 
-  if [ $askPass = 0 ]; then
-    sshCmd=(ssh -oStrictHostKeyChecking=no "$nodeAddr" /bin/bash)
-  else
-    sshCmd=(sshpass -p "$sshPass" ssh -oStrictHostKeyChecking=no "$sshUser"@"$nodeAddr" /bin/bash)
-  fi
-
-"${sshCmd[@]}" << "EOF"
-# set -o xtrace
-
-containerIds=$(docker ps -a -f name=openmpf_ -q)
-
-if [ ! -z "$containerIds" ]; then
-  echo "Removing openmpf containers:"
-  docker container rm -f $containerIds;
-else \
-  echo "No openmpf containers running."
-fi
+# Remove the shared data before the shared volume is removed.
 echo
-
-# Don't remove openmpf_shared_data.
-volumeIds=$(docker volume ls -f name=openmpf -q | sed "s/openmpf_shared_data//g")
-
-if [ ! -z "$volumeIds" ]; then
-  echo "Removing openmpf volumes:"
-  docker volume rm -f $volumeIds;
+if [ "$removeSharedData" == 1 ]; then
+  # Remove everything.
+  sh "$scriptPath" || exit -1 # abort if script fails
 else
-  echo "No openmpf volumes found."
+  # Minimally, remove the "nodes" directory.
+  sh "$scriptPath" "nodes" || exit -1 # abort if script fails
 fi
 echo
 
-exit
-EOF
-
-  echo
-done <<< "$nodeIds"
-
-# TODO: Automate this, if possible.
-echo
-echo "IMPORTANT: Please manually remove the contents of the shared (NFS) volume."
+if [ "$removeAllVolumes" == 1 ]; then
+  forAllNodes "$nodeIds" cleanupAllVolumes
+elif [ "$removeMysqlVolume" == 1 ]; then
+  forAllNodes "$nodeIds" cleanupMysqlVolume
+fi
